feat(r-toast): move manager and plugin to the plugins folder
