+++ packages/recomponents/src/components/r-toast/r-toast.spec.js
@@ -1,7 +1,7 @@
 import {createLocalVue, shallowMount} from '@vue/test-utils';
 import {renderToString} from '@vue/server-test-utils';
 import RToast from './r-toast.vue';
-import RToastPlugin from '../r-toast';
+import RToastPlugin from '../../plugins/r-toast-manager';
 
 describe('r-toast.vue', () => {
     const props = {
@@ -16,7 +16,7 @@ describe('r-toast.vue', () => {
     it('should render Wrapper and match snapshot', async () => {
         // const wrapper = localVue.$toast.positive(props.message, {autoHide: false});
         const wrapper = shallowMount(RToast, {
-            props
+            props,
         });
         expect(wrapper).toMatchSnapshot();
     });
@@ -42,16 +42,17 @@ describe('r-toast.vue', () => {
     });
 
     it('should create toasts via toast-manager', async () => {
-
         const message = `This is a manager's toast!`;
         localVue.$toast.positive(message, {autoHide: false});
         localVue.$toast.negative(message, {autoHide: false});
         localVue.$toast.warning(message, {autoHide: false});
         localVue.$toast.info(message, {autoHide: false});
         const managerToasts = [...document.getElementsByClassName(`r-toast`)]
-            .map(toast => toast['__vue__'])
-            .filter(toast => toast['_props'] && toast['_props'].message === message);
-        expect(managerToasts).toHaveLength(4);
+            /* eslint-disable no-underscore-dangle */
+            .map(toast => toast.__vue__)
+            .filter(toast => toast._props && toast._props.message === message);
+            /* eslint-enable no-underscore-dangle */
 
+        expect(managerToasts).toHaveLength(4);
     });
 });
