+++ packages/recomponents/src/plugins/r-toast-manager/index.js
@@ -1,8 +1,8 @@
-import RToast from './r-toast.vue';
+import RToast from '../../components/r-toast/r-toast.vue';
 import RToastManager from './r-toast-manager';
 
 const RToastPlugin = (Vue, options = {}) => {
-    let methods = RToastManager(Vue, options);
+    const methods = RToastManager(Vue, options);
     Vue.$toast = methods;
     Vue.prototype.$toast = methods;
 };
