+++ packages/recomponents/src/plugins/r-toast-manager/r-toast-types.js
@@ -1,4 +1,4 @@
-export const RToastTypes = {
+export default {
     Positive: 'positive',
     Negative: 'negative',
     Warning: 'warning',
